Remove Jekyll, Webpack and Docker from dependancies
