--- .browserslistrc
@@ -1,9 +1,0 @@
-# Browsers that we support
-# run `npx browserslist` for full list
-
-defaults
-not IE 11
-not IE_Mob 11
-not op_mini all
-# maintained node versions
-
