--- Gemfile
@@ -1,3 +1,0 @@
-source 'https://rubygems.org'
-gem 'github-pages', group: :jekyll_plugins
-gem 'jekyll-babel'
