--- _config.yml
@@ -1,24 +1,0 @@
-title: The Cowin Map
-description: Senior Ruby Engineer, Munich, Germany, Planet Earth
-description: "#verified #covid resources from volunteers across #India on Google maps for quick, updated and easy access."
-url: https://cowinmap.com
-
-plugins:
-  - jekyll-babel
-
-keep_files:
-  - "*.js.map"
-
-exclude:
-  - _dist
-  - node_modules
-  - assets/js/src/
-  - assets/js/dist/README.md
-  - package.json
-  - webpack.config.js
-  - README.md
-  - Gemfile
-  - Gemfile.lock
-  - CNAME
-  - docker-compose.yml
-  - Dockerfile
