--- _includes/filters/hospitals.html
@@ -1,42 +1,0 @@
-<div class="hospital-filters">
-  <div class="dropdown-item">
-    <label class="checkbox">
-      <input type="checkbox"
-             name="filters[shouldHaveCovidBeds]"
-             data-filter-name="shouldHaveCovidBeds"
-             value="1"
-             checked>
-      Covid Beds
-    </label>
-  </div>
-  <div class="dropdown-item">
-    <label class="checkbox">
-      <input type="checkbox"
-             name="filters[shouldHaveOxygenBeds]"
-             data-filter-name="shouldHaveOxygenBeds"
-             value="1"
-             checked>
-      Oxygen Beds
-    </label>
-  </div>
-  <div class="dropdown-item">
-    <label class="checkbox">
-      <input type="checkbox"
-             name="filters[shouldHaveICUs]"
-             data-filter-name="shouldHaveICUs"
-             value="1"
-             checked>
-      ICUs
-    </label>
-  </div>
-  <div class="dropdown-item">
-    <label class="checkbox">
-      <input type="checkbox"
-             name="filters[shouldHaveVentilators]"
-             data-filter-name="shouldHaveVentilators"
-             value="1"
-             checked>
-      Ventilators
-    </label>
-  </div>
-</div>
