--- _includes/filters/oxygen.html
@@ -1,42 +1,0 @@
-<div class="oxygen-filters">
-  <div class="dropdown-item">
-    <label class="checkbox">
-      <input type="checkbox"
-             name="filters[shouldHaveRefilling]"
-             data-filter-name="shouldHaveRefilling"
-             value="1"
-             checked>
-      Oxygen Refilling
-    </label>
-  </div>
-  <div class="dropdown-item">
-    <label class="checkbox">
-      <input type="checkbox"
-             name="filters[shouldHaveCylinders]"
-             data-filter-name="shouldHaveCylinders"
-             value="1"
-             checked>
-      Oxygen Cylinders
-    </label>
-  </div>
-  <div class="dropdown-item">
-    <label class="checkbox">
-      <input type="checkbox"
-             name="filters[shouldHaveCans]"
-             data-filter-name="shouldHaveCans"
-             value="1"
-             checked>
-      Oxygen Cans
-    </label>
-  </div>
-  <div class="dropdown-item">
-    <label class="checkbox">
-      <input type="checkbox"
-             name="filters[shouldHaveConcentrator]"
-             data-filter-name="shouldHaveConcentrator"
-             value="1"
-             checked>
-      Oxygen Concentrators
-    </label>
-  </div>
-</div>
