--- _layouts/default.html
@@ -1,56 +1,0 @@
-<!DOCTYPE html>
-<html lang="{{ site.lang | default: "en-US" }}">
-  <head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-
-    <script src="https://kit.fontawesome.com/d14f08e7fd.js" crossorigin="anonymous"></script>
-    <!-- Google Tag Manager -->
-    <script>(function(w,d,s,l,i){w[l]=w[l]||[];w[l].push({'gtm.start':
-      new Date().getTime(),event:'gtm.js'});var f=d.getElementsByTagName(s)[0],
-        j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src=
-'https://www.googletagmanager.com/gtm.js?id='+i+dl;f.parentNode.insertBefore(j,f);
-    })(window,document,'script','dataLayer','GTM-5KDBBZ6');</script>
-    <!-- End Google Tag Manager -->
-
-    {% seo %}
-    <link rel="stylesheet" href="{{ "/assets/css/style.css?v=" | append: site.github.build_revision | relative_url }}">
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bulma@0.9.2/css/bulma.min.css">
-  </head>
-
-  <body class="">
-    <!-- Google Tag Manager (noscript) -->
-    <noscript><iframe src="https://www.googletagmanager.com/ns.html?id=GTM-5KDBBZ6"
-                      height="0" width="0" style="display:none;visibility:hidden"></iframe></noscript>
-    <!-- End Google Tag Manager (noscript) -->
-
-    {% if site.title and site.title != page.title and page.show_header != false %}
-      <header>
-        <div class="container">
-          <div class="content">
-            <span class="is-hidden-mobile">
-              <!-- Shown on all screen sizes except mobile -->
-              <h3 class="title is-3">
-                <a href="{{ "/" | absolute_url }}">{{ site.title }}</a>
-              </h3>
-              <p>#verified #covid resources from volunteers across #India on Google maps for quick, updated and easy access.</p>
-              <p>Let’s help India win against covid!</p>
-            </span>
-
-            <span class="is-hidden-tablet">
-              <!-- Shown only on mobile -->
-              <h6 class="title is-5">
-                <a href="{{ "/" | absolute_url }}">{{ site.title }}</a>
-              </h6>
-              <p class="is-size-7">#verified #covid resources from volunteers across #India on Google maps for quick, updated and easy access.</p>
-              <p class="is-size-7">Let’s help India win against covid!</p>
-            </span>
-          </div>
-        </div>
-      </header>
-      {% endif %}
-
-      {{ content }}
-  </body>
-</html>
