--- assets/css/style.scss
@@ -1,4 +1,0 @@
----
----
-
-@import "map";
