+++ assets/js/index.js
@@ -1,8 +1,3 @@
-import City from '../city'
-import CowinMap from '../cowinMap'
-import '../bulmaInteractions'
-import '../selectAll'
-
 function initialiseMap() {
   let cowinMap = new CowinMap();
 
@@ -302,6 +297,5 @@ function initialiseMap() {
     afterCityDataFetchCallback
   );
 }
-
 
 window.initialiseMap = initialiseMap;
