+++ assets/js/lib/locations/covidBedLocation.js
@@ -28,5 +28,3 @@ class CovidBedLocation {
     return true;
   }
 }
-
-export default CovidBedLocation;
