+++ assets/js/lib/locations/oxygenSupplyLocation.js
@@ -26,5 +26,3 @@ class OxygenSupplyLocation {
     return undefined;
   }
 }
-
-export default OxygenSupplyLocation;
