--- assets/js/src/README.md
@@ -1,10 +1,0 @@
-# JS Source files
-
-All JS source files are to be in this directory.
-These are the files that developers edit.
-
-We use [Webpack](https://webpack.js.org) to bundle all these Javascript
-files into final Javascript bundles that are included in the HTML.
-
-Webpack is configured to take the JS files from `entrypoints` directory,
-bundles them as individual files and put them in the `dist` directory.
