--- docker-compose.yml
@@ -1,37 +1,0 @@
-version: '3'
-services:
-  site: &site
-    build:
-      context: ./
-    networks:
-      - cowin_dev
-    expose:
-      - 4000
-    ports:
-      - "4000:4000"
-    hostname: "app_server"
-    environment: &app_env
-      BUNDLE_PATH: /app/.bundle
-      EDITOR: vi
-    volumes:
-      - app-source:/app
-    tty: true
-    stdin_open: true
-    working_dir: "/app"
-    command: 'sh -c "bundle check && bundle exec jekyll serve -H 0.0.0.0"'
-  webpack:
-    <<: *site
-    depends_on: []
-    ports: []
-    command: "npx webpack --watch"
-
-networks:
-  cowin_dev:
-    driver: bridge
-volumes:
-  app-source:
-    driver: local
-    driver_opts:
-      type: none
-      o: bind
-      device: "${PWD}"
