--- package.json
@@ -1,27 +1,0 @@
-{
-  "name": "the_cowin_map",
-  "version": "1.0.0",
-  "description": "# Development",
-  "main": "index.js",
-  "dependencies": {
-    "@babel/core": "^7.14.0",
-    "@babel/preset-env": "^7.14.0",
-    "babel-loader": "^8.2.2",
-    "glob": "^7.1.6",
-    "webpack": "^5.36.2",
-    "webpack-cli": "^4.6.0"
-  },
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/nmahesh1412/cowin_map.git"
-  },
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/nmahesh1412/cowin_map/issues"
-  },
-  "homepage": "https://github.com/nmahesh1412/cowin_map#readme"
-}
